feat: remove Dapper

+++ RinhaBackendAPI2024Q1/Utils/Extensions/TransacaoTypeExtension.cs
@@ -4,14 +4,14 @@ namespace RinhaBackendAPI2024Q1.Utils.Extensions;
 
 internal static class TransacaoTypeExtension
 { 
-    internal static int ConvertCharToIntBasedOnTransacaoType(this char transacaoType) =>
+    internal static short ConvertCharToIntBasedOnTransacaoType(this char transacaoType) =>
         transacaoType switch
         {
             'c' => 0,
             'd' => 1,
             _ => throw new TransacaoNotSupportedException()
         }; 
-    internal  static char ConvertIntToCharBasedOnTransacaoType(this int transacaoType) =>
+    internal  static char ConvertIntToCharBasedOnTransacaoType(this short transacaoType) =>
         transacaoType switch
         {
             0 => 'c',
